requirements.txt
crewai[tools]
python-dotenv>=1.0
jinja2>=3.1
pyyaml>=6.0
